+++ web/sw.js
@@ -1,4 +1,4 @@
-const BUILD_ID = "air-drow-v506-hand-runtime-loader-fix";
+const BUILD_ID = "air-drow-v507-deploy-build-order-fix";
 const CACHE_PREFIX = "air-drow-runtime-";
 const STATIC_CACHE = `${CACHE_PREFIX}${BUILD_ID}`;
 const APP_SHELL = [